# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Administrative REST API for managing the MCP Data Platform.\nEndpoints cover system info, configuration, personas, auth keys, audit logs, and knowledge management.",
        "title": "MCP Data Platform Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/admin",
    "paths": {
...
        "/auth/keys": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all API keys (key values are never exposed, only names and roles).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth Keys"
                ],
                "summary": "List auth keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.authKeyListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new API key. The key value is returned only once.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth Keys"
                ],
                "summary": "Create auth key",
                "parameters": [
                    {
                        "description": "Key definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.authKeyCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/admin.authKeyCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            }
        },
        "/auth/keys/{name}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an API key.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth Keys"
                ],
                "summary": "Delete auth key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.statusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.authKeyCreateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "CI/CD pipeline integration"
                },
                "email": {
                    "type": "string",
                    "example": "ci@example.com"
                },
                "expires_in": {
                    "description": "e.g. \"24h\", \"720h\", \"8760h\"",
                    "type": "string",
                    "example": "720h"
                },
                "name": {
                    "type": "string",
                    "example": "ci-pipeline"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "analyst"
                    ]
                }
            }
        },
        "admin.authKeyCreateResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "CI/CD pipeline integration"
                },
                "email": {
                    "type": "string",
                    "example": "ci@example.com"
                },
                "expires_at": {
                    "type": "string"
                },
                "key": {
                    "type": "string",
                    "example": "mdp_a1b2c3d4e5f6g7h8i9j0"
                },
                "name": {
                    "type": "string",
                    "example": "ci-pipeline"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "analyst"
                    ]
                },
                "warning": {
                    "type": "string",
                    "example": "Store this key securely. It will not be shown again."
                }
            }
        },
        "admin.authKeyListResponse": {
            "type": "object",
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/auth.APIKeySummary"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "admin.problemDetail": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "resource not found"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "title": {
                    "type": "string",
                    "example": "Not Found"
                },
                "type": {
                    "type": "string",
                    "example": "about:blank"
                }
            }
        },
...
        "admin.statusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "auth.APIKeySummary": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "CI/CD pipeline integration"
                },
                "email": {
                    "type": "string",
                    "example": "ci@example.com"
                },
                "expired": {
                    "type": "boolean",
                    "example": false
                },
                "expires_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "ci-pipeline"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "analyst"
                    ]
                },
                "source": {
                    "description": "\"file\", \"database\", or \"both\"",
                    "type": "string",
                    "example": "database"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}